Updates CPTPParameterizedSPAMVec to use a better parameterization.
A new parameterization is used that is more redundant but gives a
somewhat better landscape for optimizing (entire Cholesky decomp
is parameterized).  This works slightly better in practice but
still not great - even on simulated data.  It appears a rigid
positivity constraint on the prep vectors makes it easy to get
stuck in the wrong basin.