Adds special cases for qutrit basis so qutrit GST example works.
Sometimes a "dim == 1" qutrit basis is requested, which is OK, as
this should just return the identity matrix and have no labels, like
the other basis types.  This wasn't implemented yet, however, and
so using the qutrit ("qt") basis would result in assertion failures
(for dim == 3) when trying to generate a report.