Adds fiducial pair reduction results to standard gate sets.
Each of the standard gateset files (except for std2Q_XXII.py,
which hit some technical difficulties and will be added later)
now has 'global_fidPairs' (a list of 2-tuples) and
'pergerm_fidPairsDict' (a dictionary of lists of 2-tuples) members
containing the results of running FPR on the gatesets.  This is
for convenience, so we don't have to keep re-running FPR on standard
gate sets.