Reworks read_contents to avoid error in python 2.7
The open(...) builtin doesn't accept an "encoding" argument in
python 2.7, but we sometimes need to specify this to get text file
reading to work on Windows with python3.  So read_contents have
been updated to include a try/except block that only includes the
"encoding=" argument after a plain open(filename) attempt fails
with a UnicodeDecodeError.