Revamps gauge opt.  Adds analytic jac for cptp and spam penalties.
This commit updates several aspects of gauge optimization.  First,
the penalty terms used within gaugeopt_to_target are now the same
as those used within the core GST routines.  The CPpenalty,
TPpenalty, and validSpamPenalty arguments, which controlled homebrew
penalty terms, have been replaced with cptp_penalty_factor and
spam_penalty_factor, which are more simply the sum of the trace
norms of the Jamiolkowski "gate-states" and state preps respectively.
(The TPpenalty term was dropped entirely, as it seems better to restrict
GST and gauge optimization to TP gates; I don't think this penalty
was ever used.)

This allows the analytic computation of the jacobian for each of
the penalty terms (still only when using frobenius distance as the
gate and spam op metric).

It's worth noting, however that the old validSpamPenalty term
tried to control the effects (their eigenvalues in [0,1]), and
controlled the positivity of the state preps directly rather than
via the tracenorm.  This shouldn't be a big deal, but one must
remember that using the tracenorm as a proxy for the positivity
*only* makes sense then the trace is restricted to being == 1,
i.e. for the TP case.  So, while the penalty terms *can* be used
without a TP constraint on the gates, they might not work as
expected.

Also, this commit includes a slight refactoring whereby the
jacobian computation is placed within the gaugeopt_to_target
function (maybe later to be moved to the objective-function
factory function), and code was neatened up in general.

1 and 2Q jacobian checks of the gauge optimization were
performed, showing that the analytic jacobian matches the
finite difference one as well as expected, both with and
without the use of the new penalty terms.