Adds lazy-rendering capability to html reports.
A significant update to the reporting framework which builds the
html reports.  Now workspace plots and tables are rendered using
a new "htmldir" format which specifies html+js for support files
in addition to the html+js for the main document.  The
 _merge_template function within factory.py then creates these
support files.

**This means that html reports now take on a directory-tree
structure rather than a single html file**

This makes it almost always good to include the offline directory
within the tree, and the lazy loading of both workspace outputs
and tabs via AJAX calls from the main html document decrease the
load-time of reports significantly.

The main downside to this update is that most browers (Firefox
is the only currently known exception) don't allow AJAX calls in
local files, so the viewers will need to have a local webserver
of some kind up and running to view the reports... Maybe this is
ok, or maybe there are other options - still a work in progress.

Note that anyone with a python install can start a local webserver
using:

//On Mac and Linux
python -m SimpleHTTPServer

//On Windows
python -m http.server

And currently the reports seem to load fine using Firefox.