Fixes a caching bug for numpy arrays - now cache *size* as well as data.
Previously numpy's .tostring() was used to create a hashable version of
a string, but this would cause collistions between an array of shape
(1,4) and (4,1), for instance which had the same data but a different
shape.  Now the array's shape is appended to the output of tostring
so that this cache collision is avoided.