Adds "unreliable 2Q-gate gauge opt" to do_stdpractice_gst.
Now, under all built-in gauge optimization suites, if the gateset
is a 2Q gateset and it contains one of the standard 2-qubit gate
labels (or a list custom-specified by the "unreliableGates" key of
the optional advancedOptions['all'] dictionary), then an additional
branch of all other gauge optimizations is performed in which the
unreliable 2-qubit gates are weighted at 0.01 (instead of 1).

This commit also introduces the new gauge opt suite "toggleValidSpam"
which only toggles the spam penalty, leaving the gate weight fixed
at 0.001.  This is now the default, as it specifies *fewer* gauge
opts. than the other suites and with the additional unreliable-gate
gauge opt there could be a lot of gauge opts fairly easily.