Adds PDF-rendering of html report objects and links them within reports.
One side effect of this update is that WorkspaceOuput objects used
to assign themselves random IDs upon each render(...) call.  Now
they assign themselves a randomID in __init__ and this ID persists
as the default for all render calls.  This allows rendering as
'latex' to be performed after rendering in 'html' using the same
IDs, essential to linking within the HTML document.

Still in test phase, and no interface for enabling/disabling the
PDF creation.  PDF creation was disabled right before this commit,
so as not to break as much stuff.