Adds python pickle / PKL rendering and links to report + bug fixes.
Similar to the recently added support for PDF rendering & links
within HTML reports, now tables and plots can be "rendered" as
python - meaning a pickle file of either basic python objecs
(plots) or a pandas DataFrame (tables).

This commit adds the new "python.py" formatting module (parallel
to html.py and latex.py), as well as the arguments 'pdf_links'
and 'python_links' to create_general_report(...) which toggle the
PDF and PKL file generation and link creation.

Finally, this commit fixes several bugs in the PDF rendering that
were uncovered during the parallel development for the python
links.