Adds comments for what to change to make gauge-inv tables not depend on gauge-opt.
Even though tables on Gauge-Inv. tab don't depend on gauge optimization,
they technically do b/c the tables are generated using the final
(gauge-optimized) gateset.  We could replace 'gsFinal' with
'gsFinalIter' in all the places marked with "#GI", but then the
confidence region might throw an error, since it's based on the
gauge-optimized gate set.  More work needs to be done for this,
so for now we just mark the table-generations in factory.py that
shouldn't need to use "gsFinal" but can use "gsFinalIter".

Also the 'gsEP' variable in factory.py should be updated to use
the final iteration estimate rather than the gauge-optimized
gate set, as noted in the comments.