Adds 'histogram'-type to ColorBoxPlot and adds to report summary tab.
ColorBoxPlot now takes a 'typ' argument which, in addition to 'boxes'
and 'scatter' (the old formats, previously accessed by setting
the now-removed 'scatter' argument to False/True), can be set to
'histogram' to produce a histogram of the box values.

A histogram of the logL values has been added to the summary tab
of the html report, relegating the "scatter"-type plot to the
goodness-of-fit / model-violation tab.

Note: this commit also sets a lower bound of 1e-8 on the y-axis
scale of the GramMatrixBarPlot plot to improve utility.