Updates trigger_wstable_plot_creation to work with "htmldir" rendering.
Tables within HTML reports, which now use "htmldir"-type rendering,
weren't being processed corretly because the 'tableID' argument to
the JS function 'trigger_wstable_plot_creation' was assumed to be
the *outer* div of a renderred workspace table.  When rendering
"htmldir" tables, the single-table div ID is passed in as "tableID"
so the so the function didn't set the widths and heights of table
cells corretly.  Fixed now.