Adds a "gauge invariant representation" concept to report factory.
Previously, report tables & figures for gauge-invariant quantities
would still technically depend on the gauge "switch" because those
tables would use the current gauge-optimized gateset as it is more
likely to have a confidence region computed for it.  Now, a new
switched value "gsGIRep" stores a either the un-gauge-optimized
'final iteration estimate' gate set or one of the gauge-optimized
gate sets if one is found that has a confidence region factory
associated with it (and the 'final iteration estimate' doesn't).

A corresponding "criGIRep" switched value holds the confidence
region factory (if one exists) for "gsGIRep".

This update should reduce the computation and size of the reports
by eliminating the unnecessary dependency on the gauge optimization
switch in several of the report tables.