Fixes 'htmldir' table rendering error (symptom: table plots too large)
The plot creation triggering for tables was done outside an on-ready
handler when rendered with the 'htmldir' type.  This caused the table
cell widths to be computed incorrectly (jQuery's width() was called
before the table was displayed so the width() was taken from the
TD element's style rather than computed) and in turn the plot widths
as well.  When the screen size limited the size of a table, this
would cause the table plots to render too large.

Also this commit adds a "resizable" arg to trigger_wstable_plot_creation
that when "false" causes the tables to render as large as they need to,
regardless of the screen size ; but then be non-resizable.  This is
always set of false for now, and is left for future use.