Fixes "single red box" issue on plots using LinLogColormaps.
A bug in LinLogColormap caused the normalize() function to work
incorrectly when self.vmax was less than self.trans (*entire* grayscale
would get mapped to data between min & max data values).  Now
self.vmax = max(vmax, trans) so that the grayscale is always mapped
to the values between 0 (vmin) and trans, as desired.

This commit also fixes a plotly bug in scattergl whereby the
colormap argument doesn't seem to work properly.  Using the get_color
method of the Colormap class (developed for histogram plots) gave
a quick fix for this.