Adds 'resizable' and 'autosize' as arguments to create_general_report.
Under the hood, this commit adds the ability to auto-size only upon
first-load (the default, "initial", which used to be autosize=False)
or whenever the browser window size changes ("continual", previously
autosize=True) OR not at all ("none" - a new option).  The ability
to NOT auto-size (and render everything in its "native" size is
new, and was added because sometimes it's desirable to have large
tables and plots that you need to scroll to see but whose elements
remain nicely visible.  Different combinations of the 'resizable'
(a boolean) and 'autosize' (one of {"none", "initial", "continual"})
have been lightly tested.