Adds better GI-rep selection in reports & prints warnings.
Report generation now looks for both a confidence region factory
and that factory's ability to create views (using its new
'can_construct_views' method) when selecting a gauge-invariant
representative gateset for a report.

Report gen will also log messages to stdout when confidence intervals
were requested but are not being computed (because of a missing or
incapable confidence region factory).