Reconciles frobeniusdist and residuals functions - affects gauge opt!
Several somewhat significant changes to how the `frobeniusdist`
and `residuals` functions operate:

1) Previously, 'residuals' weighted the un-squared differences by the
itemWeights factors whereas 'frobeniusdist' weighted the squared-
differences.  The latter seems preferable, and so 'residuals' has
been updated to weight the residuals with the sqrt(weight) so that
(un-normalized) frobeniusdist = sqrt( sum(residuals**2) ).  A
tempoary assert statement has been added which checks this, so in
the near future frobeniusdist will just call residuals.

2) the gateWeight and spamWeight arguments of GateSet.frobeniusdist
and GateSet.residuals are gone, in favor of allowing 'gates' and
'spam' keys within the itemWeights dictionary.  This makes the use
of the itemWeights parameter uniform across gauge-optimization and
GateSet functions, which seems much nicer.

3) docstrings have been updated for 'residuals' and 'frobeniusdist'
functions in both gateset.py and gatecalc.py.

4) gauge optimization has been updated to adjust to the changes
in the 'residuals' function.