Fixes bug in LindbladParameterizedGate jacobian & adds debug stmts.
Jacobian was working with old parameterization whereby the gate
parameters corresponding to the diagonal Cholesky "L-matrix" were
squared to get the actual L-matrix diagonals.  But this is *not*
the parameterization actually used, (the params are *not* squared)
making the jacobian incorrect.  I believe this was an inadvertent
error introduced recently when CPTP spam vectors were implemented.
In any case, the bug is fixed now and the Lindblad gate jacobian
has been re-tested against finite differences and checks out OK
now.

This commit leaves commented out in the "GateSet.num_nongauge_params"
method a bunch of debugging code that was used to isolate this issue
from its symptom (weird CPTP gauge parameter counting).  This debug
code will be removed in the next commit.