Adds "display_as" argument to SpamTable and sets default to "boxes".
This commit essentially replaces the text-version SPAM matrices
in the reports with box-plots, similar to the gates.  Since the
SPAM matrices are complex (but Hermitian), they're displayed in
a bit of a strange way currently: the upper/lower triangle displays
the real/imag parts of the off diagonal elements.  This conversion
from Hermitian -> real is performed by the new ReportableQty
method "hermitian_to_real".  We'll probably want to replace this
with something better later.