Adds support to use Masonry JS, though not enabled yet.
Adds Masonry library and support for it's use within the html
report.  It currently isn't used, however, because it seems unable
to handle 1) table sizing and 2) the dynamic size changing that occurs
when a figure caption is expanded.  It's not much of a burden on the
load time, so we'll keep loading the library for now as a future option.