Fixes a couple of minor bugs in reports/notebooks.
Fixes a display() error where "global_requirejs" was still passed
as a parameter to render(...) (outdated).

Also fixes "htmldir" rendering with resizeable widgets so that the
*entire* workspace plot/table div is make resizable rather than the
single plot/table div.  This causes the resize handles to appear in
the correct place for plots again which had been broken for a little
while now.