Adds automatic processing of ".robust" estimates.
When performing gauge optimizations, when a "robust" estimate has
the same gate set as it's non-robust version (which occurs when we
just scale some of the data but don't reoptimize the likelihood),
the result is copied from the non-robust version avoiding duplicated
effort.

When generating a report, a .robust estimate is not shown unless
its gate sets differ from the non-robust version.  The confidence
region factory of a .robust estimate can & is used for a non-robust
estimate with matching gates when that non-robust estimate doesn't
have a confidence region factory of its own.

Minor update: adds print stmt showing elapsed time of entire
report generation process.