Refactors rendering code, removing '*.dir' render types.
This commit condenses the WorkspacePlot and WorkspaceTable rendering
code considerably, replacing the *.dir render types with a the
'switched_item_mode' render option.  More code is reused, and helper
functions for merging to files has be moved to the new 'merge_helpers.py'
file.  This simplifies and cleans up the rendering code, and is meant
to prepare for separate PDF reports and the ability to independently
save plot and tables as PDF/latex files.