Fixes error-generator formula and descriptive text in report.
Updates formula and description of error generator in the
GaugeVariant_errgen.html "tab" to use an inserted formula and
description based on what type of error generator was used in
the report generation.  Text also exists for the to-be-created
"logGTi" generator.

Also fixes minor bug in notebook report creation, causing a unit
test to fail.