Renames create_general_report -> create_standard_report; adds PDF report!
This commit will break backward compatibility by updating the name
of the report generation function (hopefully to be more meaningful).
It also reduces the number of arguments to create_standard_report
by collecting seldom-used arguments into a single `advancedOptions`
dictionary argument.

The meat of this commit consists of adding the ability to create a
PDF report for a Results object containing a  single estimate and
gauge optimization.  To facilitate this, the Results and Estimate
objects now have 'view(...)' methods, allowing the user to easily
create "views" into a Results or Estimate object containing only
a subset of it's estimates and/or gauge optimizations.

A latex template for the new PDF report has been added and still
needs some polishing but seems to be working.  Note that the
'brevity' argument doesn't work yet.

This commit also renames the templates/general_report folder of
html templates to templates/standard_html_report.

Tests and Notebooks still need to be updated, and so this commit
marks the beginning of a work in progress to update the pyGSTi
reports API.