Adds data compare-ability check & auto-opening for PDF reports. Fixes a bug.
Now when multiple data sets are included in a report a compatibility
check is performed prior to attempting to construct DataComparator
objects (which would raise an exception if the DataSets were incompatible).
This allows reports to be generated with incompatible data sets
(dataset comparisons just won't be computed).  Currently "compatible"
means the DataSets must have *exactly* the same keys (in the same
order) -- this could be relaxed in future enhancements.

Secondly, the 'auto_open' argument now works for PDF reports too
(web browsers can open PDFs too!).

Thirdly, the 'ID' argument of the NotApplicable class's has been added
back (Erik accidentally deleted this in a prior commit).  In all
other render(...) methods there's no need for an ID, but NotApplicable
is a special case.  This fixes a bug: report generation would fail
if a NotApplicable object needed to rendered before this fix.