Adds "combine_robust" advanced option & behavior to create_standard_report.
The boolean "combine_robust" advanced option controls whether ".robust"
estimates should automatically get combined with their non-robust
counterparts when presenting them in reports.  The default value is
True, in which case the non-robust estimate is used to compute the
model violation figures in the report EXCEPT for those on the
"Robust data scaling" tab, which repeats the figures found elsewhere
using the robust (post-scaling) estimate.  When combine_robust is
False, then only the data scaling figure is shown on the Robust
data scaling tab and all model violation figure show the post-scaled
model violation for that estimate.