Adds initial model-testing framework for analyzing non-optimized GateSets.
This commit essentially adds a number of convenience functions for
adding estimates to a results object that are not optimized - either
over the gate set's parameters or over any gauge group.  These estimates
can thus be viewed as "model tests" in which the user supplies a model
(a GateSet) and wants to test it against the data.

The 'do_model_test' function parallels 'do_long_sequence_gst'.  The
'do_stdpractice_gst' now takes a 'modelsToTest' argument which
specifies a dictionary of GateSets that can be added as estimates.
Lastly, the Results object has a new "add_model_test" member function
which can add a GateSet to test against the data and existing estimates.

As a side effect, a new TrivialGaugeGroup class is created so that
the user can explicitly specify not to gauge optimize a gate set
without getting a warning message (from default_gauge_group being
None).