Updates scoring in germ (&fiducial) selection.
Gate- and L1-penalties are now applied to the "major" part of the
score, so that setting these values > 0 (gatePenalty in ptic) has
an effect on the greedy optimization.  Previously, the number of
amplified parameters was used as the sole "major" component of the
score.  Light testing indicates the new scoring is working as
desired.