Adds 'germLengthLimits' advanced argument to do_long_sequence_gst.
This argument is also added to the more primitive std-lists
constructors 'make_lsgst_lists' and 'make_lsgst_structs', and
allows the user to limit the maximum length of specific germs.
For examples if the maxLengths argument is [1,2,4,8,16] and
germLengthLimits is { ('Gx',): 4 } then the Gx germ will only
be repeated to acheive lenghts of 1, 2, and 4 (whereas the
other germs will be repeated to achieve lengths of 1,2,4,8, and 16).

Also adds usage of germLengthLimits to one of the report-analysis
constructions in reportBaseCase.py.