Replaces "maxGermLength" with "candidateGermCounts" in generate_germs.
Due to the need to create candidate germ sets which don't contain
*every* germ up to a certain length (the list gets too long too
quickly), this commit removes the "maxGermLength" argument of
the generate_germs function and replaces it with "candidateGermCounts",
a dictionary of counts (or a special "all upto" string) for
different germ lengths, and "candidateSeed", the base seed value
used to generate a random set of germs of some length.

Tutorials and unit tests have been updated, and unit tests now
test a smaller list of candidate germs, allowing them to run
significantly faster.