Bug fixes.  New includeLGST default.
Fixes several bugs related to testing on a windows machine without
cvxpy.

Also fixes issue with 'includeLGST' argument, which was
previously "False" unless LGST was used to seed the gate set.  This
made it easy to have estimates in the same Results object with
different values for 'includeLGST' and therefore slightly different
gate sequences (in a single Results!).  Now the default for
'includeLGST' is just True, and checks are performed to ensure that
when do_long_sequence_gst appends to an existing Results that the
gate sequences match exactly.

This commit also fixes an issue with creating data-comparison
color box plots within reports, which broke when the dataset
argument was removed (b/c I though it was unneeded).  It turns
out the 'dataset' argument *is* needed, even for data-comparison
plots, but only for computing the number of spam labels.  In any
case, this is now supplied by a "reference dataset" that is currently
just the dataset of the first DataSet being compared.