Adds JSON codec capable of serializing pygsti objects.
Adds jsoncodec.py, which adds specialized encoder and decoder
functions capable of serializing pygsti objects in JSON-compatible
format.  This addition was motivated because of incompatibility,
security, and space-efficiency issues with pickle, and jsoncodec.py
provides the standard 'dump'/'dumps'/'load'/'loads' interface of
the pickle and json packages.

Unit tests also added.

One interesting implementation note: when overriding json.dump,
just specifying the 'cls' argument was insufficient to encode objects
properly (as it *was* for json.dumps).  Running the input through
the custom encoding function first seems to fix this, but the fact
that it's needed at all seems to indicate a bug in the standard
json package's implementation of 'dump'.