Adds ability to specify a dictionary of tolerances.
The `tol` argument to core MLGST and MC2GST functions can now
be given as a dict with the keys 'relx', 'relf', 'f', and 'jac' to
specify the different types of tolerances used in the ensuing
least-squares optimization algorithms.  `tol` can still be given as
a single float, in which case the 'relf' and 'jac' members are set
to that value, 'relx' = 1e-8, and 'f' = 1.0.