Refactors objects & tools, adding new "baseobjs" sub-package.
This commit shouldn't break any backward-compatibility using high-
level functions.  It consists of a refactoring of small "lesser"
objects, which are more stand-alone, into their own "baseobjs"
sub-package.  This sub-package sits at the very bottom of the
pyGSTi dependency graph and so objects in it can be used by the tools
sub-package (which solves a major issue with the prior dependency
 graph: that functions in tools/ needed to use objects in objects/
AND vice-versa).  Unit tests have been updated as needed so that
all tests should pass (and do, on Erik's laptop).