Changes LindbladParameterizedGate to use *post*-error generator.
Now the exp(L) that is parameterized by a LindbladParameterizedGate
is the *post*-error generator, so G = exp(L)*G0, instead of the
pre-error generator (used previously where G = G0*exp(L)).  This
shouldn't make any significant difference, but it seems slightly
more intuitive than the pre-error generator, so we're switching
to this convention.

This commit also changes the default error-generator type for the
reports to "logGTi" (previously it was "logTiG").