Adds 'confidence_interval_brevity' option to create_standard_report.
This new advanced option is an integer, currently defaulting to 1,
which controls how many figures are generated with error bars (when
they're available).  Currently, 1 generates EBs for all of the plots
except the ones that take a long time in 2Q analyses and 0 generates
EBs for *all* the figures.  >= 2 will generate *no* EBs, so isn't
useful yet.