Adds ComplementSPAMVec and more global-gate-param implementation.
The second significant commit toward updating the way Gate and
SPAMVec objects within a GateSet are parameterized.  This commit
adds the ComplementSPAMVec class and removes povm_identity,
identityLabel and remainderLabel references throughout much of
the GateSet/GateCalc framework.  Basic to/from vector parameterization
seems to work, and new dirty-flag processing in Gate and SPAMVec
help to keep the GateSet's _paramvec member in sync with any
changes to the local gates that aren't caught by the label-dict
__setitem__ and __delitem__ handlers.

Unit tests indicate there is something wrong with either the Map
or Matrix-based calculator (or both) - this isn't resolved yet.