Further adds to and fixes bugs in global-paramvec implementation.
The unit tests should now pass at this commit.  Significant updates
were made to fix numerous bugs/issues discovered by the tests.
Most notable is the addition of a "parent" member of Gate and
SPAMVec objects, which was justified by the need to intelligently
copy (or not) the gpindices of a Gate/SPAMVec when it is assigned
to another (different) GateSet than one it is currently in - in
this case gpindices should *not* be copied over.  Maybe in the
future the "gpindices" member will become a "link" object which
has the parent reference within it (?).

Also, adoption of the updated methods for creating a "complement"
effect have been propagated to all the standard gate set modules
and elsewhere in pygsti.  Reference to povm_identity, _remainderLabel,
and _identityLabel have been removed.