Removes uneeded __reduce__ machinery from GateSetMember objects.
Probably due to the historical path of getting GateSets and label
dicts to pickle properly, Gate and SPAMVec classes all had their
own __reduce__ methods.  These are actually unnecessary, and can
be replaced with a single __getstate__ in GateSetMember that sets
the 'parent' member to None before pickling/serializing.  Furthermore,
the OrderedMemberDict's __reduce__ function (which *is* necessary,
in order to properly initialize the OrderedDict's members) didn't
need to .copy(None) all of it's elements.

This commit just comments out the old code -- the next commit will
remove it.