Adds GateSetChild and multiple inheritance to OrderedMemberDict.
In anticipation of GateSetMember-derived objects that also are
OrderedMemberDicts, this commit adds factors out the has-a-parent
functionality from GateSetMember into a new GateSetChild (from
which GateSetMember now derives).  This allows the parent functionality
in GateSetMember and OrderedMemberDict be shared in this common
base class, which is good in terms of code reuse now but also looks
forward to when we'll need objects that inherit from both classes
(e.g. a POVM or Instrument).