Makes gpindices a property of GateSetMember and renames get_gpindices.
gpindices is now a property so that we can do additional proessing
(for nested members like POVMs and Instruments) when the parent
member is set - namely so it can set the gpindices of its children.

The old get_gpindices method was only used to get gpindices as an
array, so it's been renamed "gpindices_as_array()".