Substantial update which adds POVM and Instrument to GateSet.
This commit adds the POVM and Instrument classes, and adds dicts
which hold these objects to GateSet.  This precipitates a number
of other changes due to the new gatestring calculation philisophy/
machinery:
- POVM and Instruments are GateSetMember objects, and, like Gates
  and SPAMVecs, are vectorized with a GateSet.  They may contain
  other stuff, but these contents need not be GateSetMember objects.
- However, POVM and Instrument objects must be able to "compile"
  themselves into dicts of pure Gate and SPAMVec objects for
  the purpose of calculations (see below).  These compiled objects
  don't need to vectorize themselves but *do* need to act like
  matrices/vectors and take derivatives wrt their parameters.
- As such, the members of POVMs and Instruments, e.g.
  ComplementSPAMVec and InstrumentGate, can get away with being
  *partial* implementations of SPAMVec and Gate, respectively.
- While contained elements can have their gpindices set relative
  to their POVM or Instrument parent's parameters when contained
  within a POVM or Instrument, once representing a *compiled* object
  the SPAMVec or Gate must have its .gpindices relative to the
  parent GateSet's parameter indices.

A related but separate change (still in progress) is that gate
strings may include prep, POVM, and instrument labels.  They are
then "compiled" to an ordered dict whose keys are "raw" (gate-only)
gate strings and whose values are lists of (prep-label, povm-effect-label)
"spamTuples".  It it this compiled dictionary that is fed into
the GateCalc functions.  As such, the GateCalc classes still only
need to deal with preps, gates, and effects (but will receive the
gates and effects compiled from the Instruments and POVMs of the
parent GateSet).

This added flexibility, however, means that we cannot assume the
*same* list of spamTuples for every gate string.  This necessitated
a significant re-working of the evaluation tree and gate calculator
objects, as the leading two dimensions of the calculator-filled
quantities (indexing the common "spamLabel" and gatestring) have
been replaced by a single spam+gatestring index.  The evaluation trees
now hold index lists or slices into this leading dimension, and
MPI gathering routines have been updated to deal with the fact that
these index lists aren't necessarily slices (as they were before).

While all of the aforementioned logic has been implemented, is
hasn't been tested and therefore almost certainly doesn't work -
so this is a bad point to pull from if you're expecting things
to work.