WIP: converting spamdefs-calcs to use upgraded GateSet
Defines a more useful GateSet.compile_gatestrings function, and
converts some of the calculations which used to use the GateSet's
spamdefs dictionary to the new more-flexible-gatestring model of
computation.  There's still a significant amount of conversion to
be done - namely in core.py.

Also a small amout of bug fixing was done to get GateSet construction
to work again, and some GateSet-pickling debugging has begun.