Continued WIP: Updates/debugs calculations; testGateSets.py passes!
This commit adds a fair amount of bug fixing and unit test updating
so that all of the unit tests within testGateSets.py now pass.  These
tests don't cover the newer (multi-POVM, instruments) functionality,
which we'll need to add later.

This commit includes a fix to GateSet.hproduct so that Gates with
nonzero hessians are computed correctly (as they were in the
bulk_* routines since an upgrade a while ago).

This commit also marks a point of further clarity regarding the
roles of "compiled" POVMs and Instruments: a POVM or Instrument is
compiled to yield a set of effects or gates, and these objects must
implement "from_vector" (I thougth they didn't need to this
previously, but the map-based calcs do).  However, it can be assumed
that the order in which the obj.from_vector(...) calls are made is
the *same* as the order returned by the compile_* routines, so that,
for instance, a POVM's effects can assume that the ComplementSPAMVec
is from_vector'd last if it is always last in the dictionary returned
from POVM.compile_effects (which it is).  This assumed ordering allows
for a coordinated 'from_vector-ing' amongst objects that share some
underlying parameters.