Continued WIP: Updates & fixes report plots
Corrects gate string labelling in the "color scatterplot" plots,
and ignores duplicates in these and "color histogram" plots to that
they're sum matches that computed in the FitComparisonTable tables
(this was checked, and "Sum = ..." debug statements are left
commented out).

Adds a new TVD color boxplot to reports, along with necessary
infrastructure.