Continued WIP: adds support for un-pickling old objects.
Old Results, GateSets, and DataSets can now be un-pickled via special
cases within the __setstate__ methods of GateSet and DataSet and using
the special `pygsti.objects.results.enable_old_python_results_unpickling`
function.  Not very well tested, but works for several example cases.
An additional change was the addition of the default verbosity level of 1
in the VerbosityPrinter constructor, which is needed for unpickling.

Also, changes default "Rep-count" type for DataSets to float32 (was uint16)
because fractional counts are needed for accurate robust data scaling.