Continued WIP: Fixes multi-dataset writing to text format. Unit tests pass.
At this commit causes all unit tests to pass again, and marks the final
commit on what was the "feature-intermediatemeas" branch for
implementing quantum instruments, and the more generic calculation
framework (which resulted in significant updates to the structure
of the GateSet and DataSet classes).  At this point, all tutorials
have also been updated an run (example notebooks, however, might
not).