2nd attempt to fix mpi unitest hanging: run1Q_end2end -> test_run1Q_end2end
Not having the 'run1Q_end2end' unittest (in testmpiMain.py) begin with
'test' was found to be the reason the mpi tests would "hang" (3 of 4
mpi worker procs wouldn't terminate) on a laptop.  This didn't seem
to be a problem on TravisCI but maybe it is now (?) - anyway, changing
the name to 'test_run1Q_end2end' seems better all around, and might
fix the TravisCI issue (which, as it turns out can sometimes be
"fixed" by just re-running the single MPI test case/job on TravisCI
that timed out).