Removes old "measurementGates" method for intermediate measurements.
Before the massive update of GateSet and DataSet to allow for quantum
instruments and multiple POVMs, one could specify a "measurementGates"
argument to various dataset functions as a temporary way of including
intermediate measurements.  This essentially told a DataSet to add
together the counts in all the rows which differ only by differing
"measurement gates", so that a DataSet would contain separate rows for
"Gmz_0" and "Gmz_1" strings.  This functionality is superceded by the
newer & fuller support for instruments (DataSets now would contain
just a single row but with more outcomes).  The measurementGates
code was never used beyond beta-testing.