Adds 'comm' argument to generate_fake_data; fixes MPI test.
A MPI unit test was occasionally hanging (and erroring TravisCI)
because the unit test used generate_fake_data separately on each
processor and in 50/50 probabilty cases numpy's 'binomial'
function seems to randomly choose which output count to be the
larger one (e.g. 480,520 vs. 520,480) resulting in different
datasets on different processors.  This resulted in different LGST
gate sets on different processors which is checked for and causes
an assert failure (but only on *some* of the procs) - hench the
hanging.  The generate_fake_data function now takes a comm argument
which currenlty doesn't speed anything up but does ensure that
each processor gets the *same* generated gate set.  The problematic
unit test now passes its comm to generate_fake_data.