Adds rudimentary "synthetic idle" decomp/recomp of error rates.
A work in progress.  Adds a "robust" option to ErrgenTable and
corresponding robust error generator projections in reportables.py
that does all the heavy lifting.  Still in a debugging state, and
given our current understanding that not all error generator terms
will be able to be re-constructed from synthetic idle sequences,
the code added here will be repurposed to a new type of linear-
combos-of-error-generator-coefficients table.