Updates TensorProdSPAMVec to have 'prep' mode too; bug fixes.
Now TensorProdSPAMVec.__init__ takes a 'typ' argument that can be
"prep" or "effect", and now can be used for a factored state
preparation vector too.

This commit also adds a number of bug fixes to the probability-
generating logic flow so that now generate_fake_data can be used
to (very slowly) generate fake data for a multi-qubit gateset
that uses many of the Map classes that have been added recently.