Optimizes sparse-matrix probability calculations; adds first Cython.
This commit updates Gate classes (in particular the GateMap-derived
classes) for faster acton(...) performance - key to computing
gate string probabilities.  It introduces the repo's first Cython
file "tools/fastcalc.pyx" which currently contains a few routines
to speed up the "embedding" action of EmbeddedGateMap and the
exp(sparse_matrix)-vector product operation used by the
LindbladParameterizedGate class.  A separate cython function is
used to compute the action of an embedded-Linblad-map that has
no unitary postfactor - which seems to be a common "special case"
worth accelerating as much as possible.

Use of the cythonized extension is fragile - I'm not sure if the
changes to setup.py will always work correctly, and the internal
use of fastcalc.py uses Cython's "pyximport" which should only be
used for development.

Final note: timing debug statements still active in gatemapcalc.py
(look for "DEBUG TIME:")